Restructure Email so that only new methods are non-static. Rename subclasses to match intent and change sendEmail functions to constructors.
