+++ class/Email/SendOIEDCertifiedNoticeEmail.php
@@ -10,7 +10,7 @@ class RegistrationConfirm extends Email{
    * @param Internship $i
    * @param Agency $agency
    */
-  public static function sendEmail(Internship $i, Agency $agency)
+  public static function __construct(Internship $i, Agency $agency)
   {
       $tpl = array();
